% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_min_r_version.R
\name{extract_min_r_version}
\alias{extract_min_r_version}
\title{Extract declared minimum R version from DESCRIPTION}
\usage{
extract_min_r_version(path = ".")
}
\arguments{
\item{path}{Path to the package root}
}
\value{
An R version number or \code{character(0)} if no minimum R version has
been declared
}
\description{
Extract declared minimum R version from DESCRIPTION
}
