% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_badges.R
\name{insert_badge_rcmdcheck}
\alias{insert_badge_rcmdcheck}
\alias{insert_badge_codecov}
\title{Insert badges in README}
\usage{
insert_badge_rcmdcheck(path = ".")

insert_badge_codecov(path = ".")
}
\arguments{
\item{path}{Path to the package root}
}
\description{
Insert badges in README
}
