% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_null_strings_linter.R
\name{na_null_strings_linter}
\alias{na_null_strings_linter}
\title{Ensure that comparisons to \code{NA} or \code{NULL} are not made as strings}
\usage{
na_null_strings_linter()
}
\description{
In an attempt to "fix" the fact that comparison to \code{NA} and \code{NULL} don't
return \code{TRUE}/\code{FALSE}, users sometimes convert them as character. This is
incorrect and will produce unexpected results.
}
\details{
This linter has been proposed for integration in the lintr package (issue
#2130) but was rejected due to the high possibility of false positives, in
the case when users actually want to compare to \code{"NA"} or \code{"NULL"} as
strings. This can happen when working with database or more generally when
the character values have not been parsed as \code{NA} or \code{NULL} by R.
}
\examples{
lintr::lint(
  text = "x == 'NA'",
  linters = etdev::na_null_strings_linter()
)

lintr::lint(
  text = "x != 'NA'",
  linters = etdev::na_null_strings_linter()
)

lintr::lint(
  text = "x \%in\% 'NA'",
  linters = etdev::na_null_strings_linter()
)

lintr::lint(
  text = "'NA' \%in\% x",
  linters = etdev::na_null_strings_linter()
)

lintr::lint(
  text = "x == 'NULL'",
  linters = etdev::na_null_strings_linter()
)

lintr::lint(
  text = "x != 'NULL'",
  linters = etdev::na_null_strings_linter()
)

lintr::lint(
  text = 'x != "NULL"',
  linters = etdev::na_null_strings_linter()
)

lintr::lint(
  text = "x \%in\% 'NULL'",
  linters = etdev::na_null_strings_linter()
)

# Pass
lintr::lint(
  text = 'x == "a"',
  linters = etdev::na_null_strings_linter()
)
}
