% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_readme.R
\name{render_readme}
\alias{render_readme}
\title{Render README.Rmd to README.md on your local developer machine.}
\usage{
render_readme()
}
\description{
Adapted from the \code{render_readme.yml} Github action.
}
\details{
Improves developer feedback and testing.
}
